# backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AssignedProfile": {
        "additionalProperties": false,
        "properties": {
          "assignmentId": {
            "format": "uuid",
            "type": "string"
          },
          "chromeProfileDir": {
            "type": "string"
          },
          "profileId": {
            "format": "uuid",
            "type": "string"
          },
          "profileName": {
            "type": "string"
          },
          "profileUsername": {
            "type": "string"
          },
          "session": {
            "nullable": true,
            "properties": {
              "errorCode": {
                "nullable": true,
                "type": "string"
              },
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "startedAt": {
                "nullable": true,
                "type": "string"
              },
              "status": {
                "enum": [
                  "LAUNCHING",
                  "ACTIVE",
                  "ERROR",
                  "CLOSED"
                ],
                "type": "string"
              }
            },
            "required": [
              "id",
              "status",
              "startedAt",
              "errorCode"
            ],
            "type": "object"
          },
          "shiftId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED",
              "RETIRED"
            ],
            "type": "string"
          },
          "validFrom": {
            "type": "string"
          },
          "validTo": {
            "type": "string"
          }
        },
        "required": [
          "assignmentId",
          "profileId",
          "profileName",
          "profileUsername",
          "status",
          "chromeProfileDir",
          "shiftId",
          "validFrom",
          "validTo",
          "session"
        ],
        "type": "object"
      },
      "AssignedProfileList": {
        "items": {
          "$ref": "#/components/schemas/AssignedProfile"
        },
        "type": "array"
      },
...
      "SessionCreateInput": {
        "additionalProperties": false,
        "properties": {
          "assignmentId": {
            "format": "uuid",
            "type": "string"
          },
          "chromeProfileDir": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "profileId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "profileId",
          "assignmentId",
          "chromeProfileDir"
        ],
        "type": "object"
      },
      "SessionPatchInput": {
        "additionalProperties": false,
        "properties": {
          "errorCode": {
            "maxLength": 80,
            "type": "string"
          },
          "errorDetail": {
            "maxLength": 500,
            "type": "string"
          },
          "status": {
            "enum": [
              "LAUNCHING",
              "ACTIVE",
              "ERROR",
              "CLOSED"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "accessToken": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API operacional de Agency OS",
    "title": "Agency OS API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/agent/devices/heartbeat": {
      "post": {
        "operationId": "AgentDevicesController.heartbeat",
        "parameters": [
          {
            "in": "header",
            "name": "x-device-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentDevices"
        ],
        "x-agency-device": true,
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": true,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/devices/heartbeat",
          "scope": "SELF",
          "shift": false
        },
        "x-agency-roles": [
          "OPERADOR"
        ]
      }
    },
...
    "/api/v1/agent/profiles/assigned": {
      "get": {
        "operationId": "AgentSessionsController.assigned",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignedProfileList"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentSessions"
        ],
        "x-agency-permissions": [
          "profiles.read"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.read"
            ],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "GET",
          "pagination": "BOUNDED_FILTER",
          "path": "/api/v1/agent/profiles/assigned",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
...
    "/api/v1/agent/sessions": {
      "post": {
        "operationId": "AgentSessionsController.open",
        "parameters": [
          {
            "in": "header",
            "name": "x-device-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentSessions"
        ],
        "x-agency-device": true,
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": true,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/sessions",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
    "/api/v1/agent/sessions/{id}": {
      "patch": {
        "operationId": "AgentSessionsController.update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-device-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionPatchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentSessions"
        ],
        "x-agency-device": true,
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": true,
          "idempotency": "NOT_APPLICABLE",
          "method": "PATCH",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/sessions/{id}",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
    "/api/v1/agent/sessions/{id}/close": {
      "post": {
        "operationId": "AgentSessionsController.close",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-device-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentSessions"
        ],
        "x-agency-device": true,
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": true,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/sessions/{id}/close",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
    "/api/v1/agent/sessions/prepare": {
      "post": {
        "operationId": "AgentSessionsController.prepare",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentSessions"
        ],
        "x-agency-permissions": [
          "profiles.read"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.read"
            ],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/sessions/prepare",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
    "/api/v1/assignments": {
      "get": {
        "operationId": "AssignmentsController.history",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Assignments"
        ],
        "x-agency-permissions": [
          "profiles.read"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.read"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "DIRECTOR_OPERATIVO",
            "COORDINADOR",
            "OPERADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "GET",
          "pagination": "OFFSET",
          "path": "/api/v1/assignments",
          "scope": "CREW",
          "shift": false
        }
      },
      "post": {
        "operationId": "AssignmentsController.create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Assignments"
        ],
        "x-agency-permissions": [
          "profiles.update"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.update"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "DIRECTOR_OPERATIVO",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/assignments",
          "scope": "CREW",
          "shift": false
        }
      }
    },
    "/api/v1/assignments/{id}/end": {
      "post": {
        "operationId": "AssignmentsController.end",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Assignments"
        ],
        "x-agency-permissions": [
          "profiles.update"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.update"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "DIRECTOR_OPERATIVO",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/assignments/{id}/end",
          "scope": "RESOURCE",
          "shift": false
        }
      }
    },
...
    "/api/v1/devices": {
      "get": {
        "operationId": "DevicesController.list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Devices"
        ],
        "x-agency-permissions": [
          "devices.manage"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "devices.manage"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "GET",
          "pagination": "UNBOUNDED",
          "path": "/api/v1/devices",
          "scope": "GLOBAL",
          "shift": false
        }
      },
      "post": {
        "operationId": "DevicesController.create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Devices"
        ],
        "x-agency-permissions": [
          "devices.manage"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "devices.manage"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/devices",
          "scope": "GLOBAL",
          "shift": false
        }
      }
    },
    "/api/v1/devices/{id}": {
      "get": {
        "operationId": "DevicesController.get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Devices"
        ],
        "x-agency-permissions": [
          "devices.manage"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "devices.manage"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "GET",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/devices/{id}",
          "scope": "RESOURCE",
          "shift": false
        }
      }
    },
    "/api/v1/devices/{id}/revoke": {
      "post": {
        "operationId": "DevicesController.revoke",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Devices"
        ],
        "x-agency-permissions": [
          "devices.manage"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "devices.manage"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/devices/{id}/revoke",
          "scope": "RESOURCE",
          "shift": false
        }
      }
    },
...
  },
  "servers": [],
...
}